voice_agents/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
pydub==0.25.1
ffmpeg-python==0.2.0
redis==5.0.7
rq==1.16.2
crewai==0.35.0
PyYAML==6.0.2
crewai-tools==0.2.6
# Audio/ML stack (comment out heavy ones if not needed locally)
librosa==0.10.2.post1
pydub==0.25.1
soundfile==0.12.1
torch==2.3.1
torchaudio==2.3.1
openai-whisper==20231117
pyloudnorm==0.1.1
noisereduce==3.0.2
elevenlabs==1.9.0
# Speaker diarization (HF token required)
pyannote.audio==3.1.1
# Optional heavy deps (enable as you implement tools)
# faster-whisper==1.0.0
# resemblyzer==0.1.1.dev0
# noisereduce==3.0.2
# torch==2.3.1
# torchaudio==2.3.1
# pyannote.audio==3.1.1

